:necktie: Output hemisphere outputs for any reconall run
