+++ CPAC/anat_preproc/anat_preproc.py
@@ -9,6 +9,7 @@ from nipype.interfaces.fsl import utils as fsl_utils
 from CPAC.pipeline import nipype_pipeline_engine as pe
 from CPAC.anat_preproc.ants import init_brain_extraction_wf
 from CPAC.anat_preproc.utils import create_3dskullstrip_arg_string, \
+    freesurfer_hemispheres, \
     fsl_aff_to_rigid, \
     mri_convert, \
     wb_command, \
@@ -2707,83 +2708,12 @@ def freesurfer_preproc(wf, cfg, strat_pool, pipe_num, opt=None):
             wf.connect(pick_tissue, f'{tissue}_mask', erode_tissues[tissue],
                        'in_file')
 
-    def split_hemi(multi_file):
-        lh = None
-        rh = None
-        for filepath in multi_file:
-            if 'lh.' in filepath:
-                lh = filepath
-            if 'rh.' in filepath:
-                rh = filepath
-        return (lh, rh)
-
-    split_surface = pe.Node(util.Function(input_names=['multi_file'],
-                                          output_names=['lh', 'rh'],
-                                          function=split_hemi),
-                            name=f'split_surface_{pipe_num}')
-    wf.connect(reconall, 'curv', split_surface, 'multi_file')
-    
-    split_pial = pe.Node(util.Function(input_names=['multi_file'],
-                                       output_names=['lh', 'rh'],
-                                       function=split_hemi),
-                         name=f'split_pial_{pipe_num}')
-    wf.connect(reconall, 'pial', split_pial, 'multi_file')
-    
-    split_smoothed = pe.Node(util.Function(input_names=['multi_file'],
-                                           output_names=['lh', 'rh'],
-                                           function=split_hemi),
-                             name=f'split_smoothed_{pipe_num}')
-    wf.connect(reconall, 'smoothwm', split_smoothed, 'multi_file')
-
-    split_spherical = pe.Node(util.Function(input_names=['multi_file'],
-                                            output_names=['lh', 'rh'],
-                                            function=split_hemi),
-                              name=f'split_spherical_{pipe_num}')
-    wf.connect(reconall, 'sphere', split_spherical, 'multi_file')
-    
-    split_sulcal_depth = pe.Node(util.Function(input_names=['multi_file'],
-                                               output_names=['lh', 'rh'],
-                                               function=split_hemi),
-                             name=f'split_sulcal_{pipe_num}')
-    wf.connect(reconall, 'sulc', split_sulcal_depth, 'multi_file')
-    
-    split_cortical_thick = pe.Node(util.Function(input_names=['multi_file'],
-                                                 output_names=['lh', 'rh'],
-                                                 function=split_hemi),
-                             name=f'split_cortical_thick_{pipe_num}')
-    wf.connect(reconall, 'thickness', split_cortical_thick, 'multi_file')
-    
-    split_cortical_volume = pe.Node(util.Function(input_names=['multi_file'],
-                                                  output_names=['lh', 'rh'],
-                                                  function=split_hemi),
-                             name=f'split_cortical_vol_{pipe_num}')
-    wf.connect(reconall, 'volume', split_cortical_volume, 'multi_file')
-    
-    split_white_surface = pe.Node(util.Function(input_names=['multi_file'],
-                                                output_names=['lh', 'rh'],
-                                                function=split_hemi),
-                             name=f'split_white_{pipe_num}')
-    wf.connect(reconall, 'white', split_white_surface, 'multi_file')
+    wf, hemisphere_outputs = freesurfer_hemispheres(wf, reconall, pipe_num)
 
     outputs = {
         'space-T1w_desc-brain_mask': (fill_fs_brain_mask, 'out_file'),
         'freesurfer-subject-dir': (reconall, 'subjects_dir'),
-        'lh-surface-curvature': (split_surface, 'lh'),
-        'rh-surface-curvature': (split_surface, 'rh'),
-        'lh-pial-surface-mesh': (split_pial, 'lh'),
-        'rh-pial-surface-mesh': (split_pial, 'rh'),
-        'lh-smoothed-surface-mesh': (split_smoothed, 'lh'),
-        'rh-smoothed-surface-mesh': (split_smoothed, 'rh'),
-        'lh-spherical-surface-mesh': (split_spherical, 'lh'),
-        'rh-spherical-surface-mesh': (split_spherical, 'rh'),
-        'lh-sulcal-depth-surface-map': (split_sulcal_depth, 'lh'),
-        'rh-sulcal-depth-surface-map': (split_sulcal_depth, 'rh'),
-        'lh-cortical-thickness-surface-map': (split_cortical_thick, 'lh'),
-        'rh-cortical-thickness-surface-map': (split_cortical_thick, 'rh'),
-        'lh-cortical-volume-surface-map': (split_cortical_volume, 'lh'),
-        'rh-cortical-volume-surface-map': (split_cortical_volume, 'rh'),
-        'lh-white-matter-surface-mesh': (split_white_surface, 'lh'),
-        'rh-white-matter-surface-mesh': (split_white_surface, 'rh'),
+        **hemisphere_outputs,
         'raw-average': (reconall, 'rawavg'),
         'brainmask': (reconall, 'brainmask'),
         'T1': (reconall, 'T1')
@@ -3020,6 +2950,22 @@ def freesurfer_abcd_preproc(wf, cfg, strat_pool, pipe_num, opt=None):
      "outputs": ["desc-restore_T1w",
                  "desc-restore-brain_T1w",
                  "desc-fast_biasfield",
+                 "lh-surface-curvature",
+                 "rh-surface-curvature",
+                 "lh-pial-surface-mesh",
+                 "rh-pial-surface-mesh",
+                 "lh-smoothed-surface-mesh",
+                 "rh-smoothed-surface-mesh",
+                 "lh-spherical-surface-mesh",
+                 "rh-spherical-surface-mesh",
+                 "lh-sulcal-depth-surface-map",
+                 "rh-sulcal-depth-surface-map",
+                 "lh-cortical-thickness-surface-map",
+                 "rh-cortical-thickness-surface-map",
+                 "lh-cortical-volume-surface-map",
+                 "rh-cortical-volume-surface-map",
+                 "lh-white-matter-surface-mesh",
+                 "rh-white-matter-surface-mesh",
                  "wmparc",
                  "freesurfer-subject-dir"]}
     '''
@@ -3137,12 +3083,16 @@ def freesurfer_abcd_preproc(wf, cfg, strat_pool, pipe_num, opt=None):
     wf.connect(normalize_head, 'out_file',
                reconall, 'T1_files')
 
+    wf, hemisphere_outputs = freesurfer_hemispheres(wf, reconall, pipe_num)
+
     outputs = {
         'desc-restore_T1w': (fast_correction, 'outputspec.anat_restore'),
-        'desc-restore-brain_T1w': (fast_correction, 'outputspec.anat_brain_restore'),
+        'desc-restore-brain_T1w': (fast_correction,
+                                   'outputspec.anat_brain_restore'),
         'desc-fast_biasfield': (fast_correction, 'outputspec.bias_field'),
         'wmparc': (reconall, 'wmparc'),
         'freesurfer-subject-dir': (reconall, 'subjects_dir'),
+        **hemisphere_outputs
     }
 
     return (wf, outputs)
